Replace ValidatorDelegationView with StakeDelegationView
Removed ValidatorDelegationView and introduced StakeDelegationView for displaying delegation information. Updated StakeScene and StakeDelegationViewModel to use the new view and refactored date formatting logic for delegation completion dates.
